mcp-servers/db-vector/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
chromadb>=0.5.0
psycopg2-binary>=2.9.9
sentence-transformers>=3.0.0
pydantic>=2.10.0
